+++ test/framework/easystacks/test_easystack_basic.yaml
@@ -3,4 +3,11 @@ software:
     toolchains:
       GCCcore-4.9.3:
         versions:
-          [2.25]
+          2.25:
+          2.26:
+  toy:
+    toolchains:
+      gompi-2018a:
+        versions:
+          0.0:
+            versionsuffix: '-test'
